// src/types.ts
export type Action = {
    type: string;
    payload: any;
    meta?: {
        skipHistory?: boolean;
        [key: string]: any;
    }
};
export type ToDispatch = Action[];

export interface Operation {
    add: (...actions: Action[]) => void;
    clear: () => void;
    addDiff: (fn: DiffFactoryFn) => void;
    performDiff: (fn: DiffFactoryFn) => void;
    submit: () => void;
    state: ActionState;
}

export type CardinalDirection = "n" | "w" | "s" | "e";
export type IntercardinalDirection = "ne" | "nw" | "se" | "sw";

export type HSLColor = [number, number, number];
export type RGBColor = [number, number, number];
export type RGBAColor = [number, number, number, number];

export interface LayerTransform {
    origin: Vec2;
    originBehavior: OriginBehavior;
    translate: Vec2;
    anchor: Vec2;
    rotation: number; // Radians
    scaleX: number;
    scaleY: number;
    matrix: Mat2;
}

export enum ValueType {
    Number = "number",
    Vec2 = "vec2",
    Rect = "rect",
    RGBAColor = "rgba",
    RGBColor = "rgb",
    TransformBehavior = "transform_behavior",
    OriginBehavior = "origin_behavior",
    Path = "path",
    FillRule = "fill_rule",
    LineCap = "line_cap",
    LineJoin = "line_join",
    Any = "any",
}
// ...
export type TransformBehavior = "recursive" | "absolute_for_computed";
export type OriginBehavior = "relative" | "absolute";
export type FillRule = "evenodd" | "nonzero";
export type LineCap = "butt" | "round" | "square";
export type LineJoin = "miter" | "round" | "bevel";